\name{Gierer-package}
\alias{Gierer-package}
\alias{Gierer}
\docType{package}
\title{
Overview of Gierer 2d model.
}
\description{
Implementation based on David Nissenbaum and David Sterratt's code.
}
\details{
\tabular{ll}{
Package: \tab Gierer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-31\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Stephen Eglen.
}
\references{
Gierer2d}
\examples{
source(system.file('examples/runme.R', package='gierer'))
}
